requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
manim==0.19.1
Pillow==10.2.0
boto3==1.34.0